requirements.txt
ipython
google_auth_oauthlib
google_api_python_client
folium==0.12.1.post1
matplotlib==3.3.4
numpy==1.19.2
pandas==1.1.5
protobuf==3.19.1
streamlit==1.2.0
streamlit_folium==0.4.0
